Initial DDIM setup
Set up SOME of the DDIM code, mostly from
https://github.com/ermongroup/ddim. Currently basically nothing is
working besides the UNet (see main.ipynb for how to invoke one). The
Diffusion class (which actually performs the training and whatnot) is
still being built.

Think of this commit as a backup commit in case my laptop spontaneously
combusts, which is certainly a non-zero probability.